# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: HaulBrokr API
  version: 1.0.0
  description: HaulBrokr API — Uber for dump trucks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trucks:
    get:
      operationId: listTrucks
      tags: [trucks]
      summary: List trucks (own fleet for providers, available for customers)
      parameters:
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of trucks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Truck"
    post:
      operationId: createTruck
      tags: [trucks]
      summary: Add a truck to the fleet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckInput"
      responses:
        "201":
          description: Truck created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
  /trucks/{id}:
    get:
      operationId: getTruck
      tags: [trucks]
      summary: Get a truck by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Truck found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
        "404":
          description: Truck not found
    patch:
      operationId: updateTruck
      tags: [trucks]
      summary: Update a truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruckUpdate"
      responses:
        "200":
          description: Truck updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
    delete:
      operationId: deleteTruck
      tags: [trucks]
      summary: Remove a truck from the fleet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Truck deleted

  # ── Job Requests ─────────────────────────────────────────────────────────────
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List job requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, bidding, accepted, in_progress, completed, cancelled]
        - name: mine
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of job requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Post a new job request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a job request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequest"
        "404":
          description: Request not found
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a job request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobRequestUpdate"
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequest"
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Cancel a job request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Request cancelled

  # ── Bids ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Truck:
      type: object
      required: [id, ownerId, ownerCompany, truckType, capacityTons, ratePerHour, isAvailable, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerCompany:
          type: string
        truckType:
          type: string
          enum: [standard, articulated, side_dump, bottom_dump, transfer, dump_truck, super_10, end_dump, belly_dump, lowboy, water_truck, excavator, dozer, skid_steer]
        capacityTons:
          type: number
        ratePerHour:
          type: number
        licensePlate:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        truckNumber:
          type: ["string", "null"]
        vin:
          type: ["string", "null"]
        coiStatus:
          type: string
          enum: [none, pending, active, expired]
        assignedDriverId:
          type: ["integer", "null"]
        isAvailable:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TruckInput:
      type: object
      required: [truckType, capacityTons, ratePerHour]
      properties:
        truckType:
          type: string
          enum: [standard, articulated, side_dump, bottom_dump, transfer, dump_truck, super_10, end_dump, belly_dump, lowboy, water_truck, excavator, dozer, skid_steer]
        capacityTons:
          type: number
        ratePerHour:
          type: number
        licensePlate:
          type: string
        year:
          type: integer
        make:
          type: string
        model:
          type: string
        truckNumber:
          type: string
        vin:
          type: string
        coiStatus:
          type: string
          enum: [none, pending, active, expired]
        assignedDriverId:
          type: integer
        isAvailable:
          type: boolean
        notes:
          type: string
    TruckUpdate:
      type: object
      properties:
        truckType:
          type: string
          enum: [standard, articulated, side_dump, bottom_dump, transfer, dump_truck, super_10, end_dump, belly_dump, lowboy, water_truck, excavator, dozer, skid_steer]
        capacityTons:
          type: number
        ratePerHour:
          type: number
        licensePlate:
          type: string
        year:
          type: integer
        make:
          type: string
        model:
          type: string
        truckNumber:
          type: string
        vin:
          type: string
        coiStatus:
          type: string
          enum: [none, pending, active, expired]
        assignedDriverId:
          type: integer
        isAvailable:
          type: boolean
        notes:
          type: string

    # ── Job Request ────────────────────────────────────────────────────────────
    JobRequest:
      type: object
      required: [id, customerId, customerCompany, materialType, quantityTons, pickupAddress, deliveryAddress, scheduledDate, status, trucksNeeded, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerCompany:
          type: string
        materialType:
          type: string
          enum: [dirt, gravel, sand, concrete, asphalt, demolition, topsoil, fill, other]
        quantityTons:
          type: number
        pickupAddress:
          type: string
        deliveryAddress:
          type: string
        scheduledDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [open, bidding, accepted, in_progress, completed, cancelled]
        trucksNeeded:
          type: integer
        budgetPerHour:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        bidCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    JobRequestInput:
      type: object
      required: [materialType, quantityTons, pickupAddress, deliveryAddress, scheduledDate, trucksNeeded]
      properties:
        materialType:
          type: string
          enum: [dirt, gravel, sand, concrete, asphalt, demolition, topsoil, fill, other]
        quantityTons:
          type: number
        pickupAddress:
          type: string
        deliveryAddress:
          type: string
        scheduledDate:
          type: string
          format: date-time
        trucksNeeded:
          type: integer
        budgetPerHour:
          type: number
        notes:
          type: string
    JobRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, bidding, accepted, in_progress, completed, cancelled]
        materialType:
          type: string
          enum: [dirt, gravel, sand, concrete, asphalt, demolition, topsoil, fill, other]
        quantityTons:
          type: number
        pickupAddress:
          type: string
        deliveryAddress:
          type: string
        scheduledDate:
          type: string
          format: date-time
        trucksNeeded:
          type: integer
        budgetPerHour:
          type: number
        notes:
          type: string

    # ── Bid ───────────────────────────────────────────────────────────────────
...